pyproject.toml
[project]
name = "Document-Image-Translator"
version = "0.1.0"
description = "Translates document images from source to target languages"
readme = "README.md"
requires-python = ">=3.12.4"
authors = [
    { name = "Temiloluwa Adeoti", email = "temmiecvml@gmail.com" },
]

[dependency-groups]
prod = [
    "aioboto3>=14.3.0",
    "aiofiles>=24.1.0",
    "beautifulsoup4>=4.13.4",
    "datauri>=1.0.0",
    "google-genai>=1.13.0",
    "mistralai>=1.7.0",
    "omegaconf>=2.3.0",
    "pillow>=11.2.1",
    "python-json-logger>=3.3.0",
    "requests>=2.32.3",
]

lint = [
    "mypy>=1.15.0",
    "pre-commit>=4.2.0",
    "ruff>=0.11.7",
]
tests = [
    "pytest>=8.3.5",
    "pytest-cov>=6.1.1",
    "numpy>=1.26.4",
    "pytest-asyncio>=1.0.0",
    "pytest-dotenv>=0.5.2",
]
api = [
    "python-ulid>=3.0.0",
]

[tool.uv]
default-groups = ["prod"]

[tool.mypy]
python_version = 3.12
warn_unused_configs = true
warn_return_any = true
warn_unused_ignores = true
warn_unreachable = true
check_untyped_defs = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
